feat: add jimple resources
